+++ db/cql_type_parser.hh
@@ -49,7 +49,10 @@
 #include "seastarx.hh"
 
 class types_metadata;
+
+namespace data_dictionary {
 class keyspace_metadata;
+}
 
 namespace db {
 namespace cql_type_parser {
@@ -58,7 +61,7 @@ data_type parse(const sstring& keyspace, const sstring& type);
 
 class raw_builder {
 public:
-    raw_builder(keyspace_metadata &ks);
+    raw_builder(data_dictionary::keyspace_metadata &ks);
     ~raw_builder();
 
     void add(sstring name, std::vector<sstring> field_names, std::vector<sstring> field_types);
